requirements.txt
streamlit>=1.30.0
openai>=1.10.0
paddleocr>=2.6.1
paddlepaddle>=2.5.0
pillow>=10.0.0
numpy>=1.23.0